requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
Django==3.1.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
environs==9.2.0
gunicorn==19.9.0
idna==2.10
jieba==0.42.1
marshmallow==3.9.1
mysqlclient==2.0.1
Naked==0.1.31
numpy==1.19.3
Pillow==8.0.1
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
PyMySQL==0.10.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
channels==3.0.2
channels_redis==3.2.0
celery==5.0.5
redis==3.5.3